frontend/ai-server/requirements.txt
fastapi
uvicorn
rembg
pillow
numpy
opencv-python
python-multipart